Move login from controller into module
